api.pubsub: move `Subscription` model to `api.models`
Move `Subscription` model definition to module
containing all the generic API models i.e.
`api/models.py` to keep coding style consistent.
